refactor: Move logic for compiling shorebird.yaml into a new file
I also changed the logic from using a StringBuilder to using
YamlEdit, that might change the output order of the yaml file
but should *not* change the contents.

I also added our first test of this logic.

Next step is to make a `flutter shorebird compile-yaml` command
or similar, and then have flutter.groovy call that so that we can
share this logic between platforms and have a single place to
edit when we need to make changes.